# Conflicts:
#	Sail/src/API/DX12/resources/DX12Texture.cpp
#	Sail/src/Sail/events/Events.h

+++ Sail/src/Network/NetworkStructs.hpp
@@ -5,7 +5,7 @@ const unsigned int HOST_META_DESC_SIZE = MAX_PACKAGE_SIZE - 6;
 
 // The length of the string you get from archiving an int.
 // This is needed to know how many bytes to read to find the size of a message.
-constexpr int MSG_SIZE_STR_LEN = 1 + sizeof(int);
+constexpr size_t MSG_SIZE_STR_LEN = 1 + sizeof(size_t);
 
 //#define DEBUG_NETWORK
